refactor : 구현체들의 이름 변경
