+++ src/test/java/domain/FixedDeck.java
@@ -1,14 +1,14 @@
 package domain;
 
 import domain.card.Card;
-import domain.deck.CardDeck;
+import domain.deck.Deck;
 import java.util.List;
 
-public class StubDeck implements CardDeck {
+public class FixedDeck implements Deck {
     private final List<Card> deck;
     private int index;
 
-    public StubDeckImplTest(List<Card> deck) {
+    public FixedDeck(List<Card> deck) {
         this.deck = deck;
         this.index = 0;
     }
